pvscbot/requirements.txt
aiohttp==3.4.4
appdirs==1.4.3
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
black==18.9b0
chardet==3.0.4
Click==7.0
colorama==0.4.0
gidgethub==3.0.0
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.7
idna-ssl==1.1.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.5.1
protobuf==3.6.1
ptvsd==4.2.0
pylint==2.1.1
six==1.11.0
toml==0.10.0
typed-ast==1.1.0
uritemplate==3.0.0
wrapt==1.10.11
yarl==1.2.6
